api-gateway: separation between port-in and business logic
